+++ magik-checks/src/test/java/org/stevenlooman/sw/magik/checks/FormattingCheckTest.java
@@ -6,6 +6,9 @@ import org.junit.Test;
 import org.stevenlooman.sw.magik.MagikCheck;
 import org.stevenlooman.sw.magik.MagikIssue;
 
+import java.io.IOException;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.List;
 
 public class FormattingCheckTest extends MagikCheckTestBase {
@@ -168,6 +171,46 @@ public class FormattingCheckTest extends MagikCheckTestBase {
     String code = "_pragma(classify_level=restricted, topic={a,b})";
     List<MagikIssue> issues = runCheck(code, check);
     assertThat(issues).isEmpty();
+  }
+
+  @Test
+  public void testSyntaxError() {
+    MagikCheck check = new FormattingCheck();
+    String code = "_method";
+    List<MagikIssue> issues = runCheck(code, check);
+    assertThat(issues).isEmpty();
+  }
+
+  @Test
+  public void testFile() throws IllegalArgumentException, IOException {
+    MagikCheck check = new FormattingCheck();
+    Path path = Paths.get("magik-checks/src/test/resources/test_product/test_module/source/in_load_list.magik");
+    List<MagikIssue> issues = runCheck(path, check);
+    assertThat(issues).isEmpty();
+  }
+
+  @Test
+  public void testLineStartWithTabs() {
+    MagikCheck check = new FormattingCheck();
+    String code = "\tprint(a)";
+    List<MagikIssue> issues = runCheck(code, check);
+    assertThat(issues).isEmpty();
+  }
+
+  @Test
+  public void testLineStartWithSpaces() {
+    MagikCheck check = new FormattingCheck();
+    String code = "        print(a)";
+    List<MagikIssue> issues = runCheck(code, check);
+    assertThat(issues).isNotEmpty();
+  }
+
+  @Test
+  public void testLineStartWithTabsSpaces() {
+    MagikCheck check = new FormattingCheck();
+    String code = " \tprint(a)";
+    List<MagikIssue> issues = runCheck(code, check);
+    assertThat(issues).isNotEmpty();
   }
 
 }